Obter heroi por id

+++ Backend/src/Supers.Infrastructure/Dados/Repositorio/SuperHeroiRepository.cs
@@ -18,34 +18,29 @@ namespace Supers.Infrastructure.Dados.Repositorio
             return await _dbContext.SuperHerois.Include(h => h.HeroisSuperPoderes).ThenInclude(hsp => hsp.SuperPoderes).ToListAsync();
         }
 
-        public async Task<SuperHeroi> ObterHeroiPorId(int id)
+        public async Task<SuperHeroi?> ObterHeroiPorId(int id)
         {
-            return await _dbContext.SuperHerois.AsNoTracking().FirstOrDefaultAsync(x => x.Id == id);
+            return await _dbContext.SuperHerois
+                .Include(heroi => heroi.HeroisSuperPoderes)
+                    .ThenInclude(relacionamento => relacionamento.SuperPoderes) 
+                .FirstOrDefaultAsync(heroi => heroi.Id == id);
         }
 
-        public async Task AtualizarHeroiPorId(int id, SuperHeroi heroi)
+        public Task AtualizarHeroiPorId(int id, SuperHeroi heroi)
         {
-            var heroiASerAtualizado = await ObterHeroiPorId(id);
-
-            if (heroiASerAtualizado != null)
-            {
-                heroiASerAtualizado.Nome = heroi.Nome;
-                heroiASerAtualizado.NomeHeroi = heroi.NomeHeroi;
-                heroiASerAtualizado.DataNascimento = heroi.DataNascimento;
-                heroiASerAtualizado.Altura = heroi.Altura;
-                heroiASerAtualizado.Peso = heroi.Peso;
-                heroiASerAtualizado.HeroisSuperPoderes = heroi.HeroisSuperPoderes;
-
-                _dbContext.SuperHerois.Update(heroiASerAtualizado);
-                await _dbContext.SaveChangesAsync();
-            }
+            _dbContext.SuperHerois.Update(heroi);
+            return Task.CompletedTask;
         }
 
         public async Task ExcluirHeroiPorId(int id)
         {
             var heroiASerExcluido = await ObterHeroiPorId(id);
 
-            _dbContext.SuperHerois.Remove(heroiASerExcluido);
+            if (heroiASerExcluido != null)
+            {
+                _dbContext.SuperHerois.Remove(heroiASerExcluido);
+            };
+            
         }
 
         public async Task<bool> ExisteHeroiCadastradoPorNomeHeroi(string nomeHeroi) => await _dbContext.SuperHerois.AnyAsync(x => x.NomeHeroi.Equals(nomeHeroi));
